requirements.txt
deepgram-sdk>=3.0.0
pyaudio>=0.2.11
asyncio-mqtt>=0.16.0
python-dotenv>=1.0.0
colorama>=0.4.6
numpy>=1.21.0
websockets>=11.0.0
openai>=1.0.0
requests>=2.25.0
pyyaml>=6.0.0 